refactor: CourseItemListener 중복 코드 제거
