+++ android/app/src/main/java/io/coursepick/coursepick/presentation/course/ExploreCoursesFragment.kt
@@ -1,22 +1,19 @@
 package io.coursepick.coursepick.presentation.course
 
-import android.Manifest
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.annotation.RequiresPermission
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import io.coursepick.coursepick.databinding.FragmentExploreCoursesBinding
-import io.coursepick.coursepick.presentation.Logger
 
 class ExploreCoursesFragment : Fragment() {
     @Suppress("ktlint:standard:backing-property-naming")
     private var _binding: FragmentExploreCoursesBinding? = null
     private val binding get() = _binding!!
-    private val courseAdapter by lazy { CourseAdapter(CourseItemListener()) }
     private val viewModel: CoursesViewModel by activityViewModels { CoursesViewModel.Factory }
+    private val courseAdapter by lazy { CourseAdapter(DefaultCourseItemListener(viewModel)) }
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -51,29 +48,4 @@ class ExploreCoursesFragment : Fragment() {
             courseAdapter.submitList(state.courses)
         }
     }
-
-    private fun CourseItemListener(): CourseItemListener =
-        object : CourseItemListener {
-            override fun select(course: CourseItem) {
-                Logger.log(
-                    Logger.Event.Click("course_on_list"),
-                    "id" to course.id,
-                    "name" to course.name,
-                )
-                viewModel.select(course)
-            }
-
-            override fun toggleFavorite(course: CourseItem) {
-                viewModel.toggleFavorite(course)
-            }
-
-            @RequiresPermission(anyOf = [Manifest.permission.ACCESS_FINE_LOCATION, Manifest.permission.ACCESS_COARSE_LOCATION])
-            override fun navigateToMap(course: CourseItem) {
-                Logger.log(
-                    Logger.Event.Click("navigate"),
-                    "id" to course.id,
-                    "name" to course.name,
-                )
-            }
-        }
 }
